requirements.txt
PyQt5>=5.9.2
opencv-python-headless>=4.5.1.48
pyqtgraph>=0.11.1
pymssql>=2.2.2
requests>=2.22.0
qdarkstyle>=2.8.1
python-dotenv>=0.19.0